Move text parser to separate module
